Update to the scene system

+++ src/input/Controller.h
@@ -15,7 +15,7 @@ namespace Atlas {
 			ControllerHandler(Camera* camera, float sensibility, float speed, float reactivity,
 					float threshold, int32_t device = -1);
 
-			void Update(Camera* camera, uint32_t deltaTime);
+			void Update(Camera* camera, float deltaTime);
 
 			float sensibility;
 			float speed;
